Formulario de Producto en ultima fase falta el update definitivo

+++ system/libraries/Session/drivers/Session_files_driver.php
@@ -368,7 +368,13 @@ class CI_Session_files_driver extends CI_Session_driver implements SessionHandle
 				continue;
 			}
 
-			unlink($this->_config['save_path'].DIRECTORY_SEPARATOR.$file);
+			//unlink($this->_config['save_path'].DIRECTORY_SEPARATOR.$file);
+
+			if( is_file($this->_config['save_path'].DIRECTORY_SEPARATOR.$file) )
+			{
+				unlink($this->_config['save_path'].DIRECTORY_SEPARATOR.$file);
+			}
+			
 		}
 
 		closedir($directory);
